Simplify the implementation of the crate and fix #1 at the same time.
Drops the `paste` crate and does not bother to use `#[wasm_bindgen]` to
generate the exported init function. Instead, exports a plain `extern
"C"` function and uses uses `#[export = name]` to set its name, as
`wasm-bindgen` would have done.

This is both simpler, and also does not require that the calling crate
have `wasm_bindgen` as a direct dependency, so it fixes #1 at the same
time.